requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.77.0
decorator==5.2.1
-e git+https://github.com/MNGARCIA085/DeepChest.git@b0fd9bc157987137da1b93df3f4c83b5bf6b8053#egg=deep_chest
defusedxml==0.7.1
docker==7.1.0
docopt==0.6.2
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.24.3
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-pasta==0.2.0
gradio==6.6.0
gradio_client==2.1.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
groovy==0.1.2
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.4.1
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
namex==0.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.6
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.0
pip-chill==1.0.3
pipreqs==0.5.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.0.1
yarg==0.1.9
zipp==3.23.0
